+++ src/components/Header/Header.js
@@ -11,6 +11,7 @@ import { Menu } from "../Menu/Menu";
 import { Image } from "../Image/Image";
 import EN from "../../translations/en.json";
 import logoFile from "../../assets/sig-blk-en.svg";
+import logoFilefr from "../../assets/sig-blk-fr.svg";
 import { SearchBar } from "../SearchBar/SearchBar";
 import { Language } from "../Language/Language";
 import { Breadcrumb } from "../Breadcrumb/Breadcrumb";
@@ -24,22 +25,22 @@ export function Header(props) {
     <div className="ds-header" id={props.id}>
       <header>
         <div className="ds-px-4 sm:ds-container md:ds-flex-nowrap md:ds-flex-row ds-flex ds-flex-wrap ds-justify-between ds-items-center ds-mx-auto">
-          <a
-            href="#"
-            className={`ds-flex ds-w-20px ds-h-20px ${props.className}`}
-          >
-            <Image src={props.logo} alt={props.altText} />
+          <a href="#" className={`${props.className}`}>
+            <Image
+              src={props.lng === "fr" ? { logoFilefr } : props.logo}
+              alt={props.altText}
+            />
           </a>
           {/* Developer Note: This will be moved as seperate component once language translater component is implemented */}
           <section className="ds-flex md:ds-hidden">
             <Language id="lang2" />
           </section>
           {/* Developer Note: This will be moved as seperate component once search component is implemented */}
-          <section className="ds-w-full md:ds-flex md:ds-w-284px ds-py-2">
+          <section className="ds-w-full md:ds-flex md:ds-w-332px ds-py-2">
             <SearchBar onChange={onChange} onSubmit={onSubmit} />
           </section>
           {/* Developer Note: This will be moved as seperate component once language translater component is implemented */}
-          <section className="ds-hidden md:ds-flex ds-pl-4 ds-pr-8">
+          <section className="ds-hidden md:ds-flex ds-pl-4 ds-pr-8 md:ds-pr-0">
             <Language id="lang1" />
           </section>
         </div>